Attempt moving 1.8 files to a separate module
