--- resources/config.yml
@@ -1,19 +1,0 @@
-config-version: 2
-currency: '$'
-shop-perm: customshop.use
-admin-perm: customshop.admin
-chestshop-perm: customshop.chestshop
-items-sold-message: '&a <amount> item(s) sold for <currency><price>'
-item-buy-lore: '&cLeftClick &7to buy &c<amount> &7for &c<currency><price>'
-stack-buy-lore: '&6Shift+LeftClick &7to buy &6<amount> &7for &c<currency><price>'
-item-sell-lore: '&cRightClick &7to sell &c<amount> &7for &c<currency><price>'
-stack-sell-lore: '&6Shift+RightClick &7to sell &c<amount> &7for &c<currency><price>'
-no-permission-message: '&cYou are not allowed to use this command!'
-update-found: '&e&lCustom Shop is ready to update with &6&l/shop update'
-spawner-name: '&a<spawner> Spawner'
-spawner-placed-message: '&7<spawner> placed'
-#These value should be unique to CustomShop
-shop-inventory-title: '&7CS Inventory'
-sell-inventory-title: '&7CS Sell'
-#Should players be able to auto-buy stacks of items?
-enable-stack-buying: true
