--- resources/shops.yml
@@ -1,149 +1,0 @@
-shops:
- ores:
-  name: '&cOres'
-  permission: 'customshop.main'
-  items:
-  #List the items you want in this format
-  # ID:DATA, AMOUNT_BUY, PRICE_BUY, AMOUNT_SELL, PRICE_SELL, buy:[COMMAND</>COMMAND], sell:[COMMAND</>COMMAND]
-  #Put a - in place of anything that you want to disable.
-  #NOTE: You don't have to add commands. All commands are executed from console.
-  #      The command will be executed for each batch sold. Eg) buy commands will be executed twice if a player buys
-  #       2x (13 Glowstone Blocks).
-  #      You can also use variables such as <player>, <price>, <amount> etc.
-   quartz: '406:0,16,200,0,-'
-   lapis: '351:4,16,200,1,2'
-   coal: '263:0,16,100,1,2'
-   iron: '265:0,16,2000,1,7'
-   gold: '266:0,16,2400,1,35'
-   emerald: '388:0,16,12000,1,85'
-   diamond: '264:0,16,4800,1,65'
-   quartzblock: '155:0,64,600,1,1'
-   redstoneblock: '152:0,1,175,1,25'
-   lapisblock: '22:0,1,138,1,25'
-   coalblock: '173:0,1,60,1,25'
-   ironblock: '42:0,1,1125,1,90'
-   goldblock: '41:0,1,1350,1,324'
-   emeraldblock: '133:0,1,6750,1,765'
-   diamondblock: '57:0,1,2780,1,585'
- food:
-  name: '&cFood'
-  permission: 'customshop.main'
-  items:
-   cookedfish: '350:0,16,150,1,2'
-   cookedchicken: '366:0,16,150,1,2'
-   cookedporkchop: '320:0,16,150,1,4'
-   cookedbeef: '364:0,16,150,1,4'
- farming:
-  name: '&cFarming'
-  permission: 'customshop.main'
-  items:
-   pumpkinseeds: '361:0,64,400,0,-'
-   cocoabeans: '351:3,64,400,0,-'
-   melonseeds: '362:0,64,400,0,-'
-   potato: '392:0,16,640,1,15'
-   carrot: '391:0,4,200,1,15'
-   netherwart: '372:0,4,2400,1,20'
-   pumpkin: '86:0,32,1200,1,18'
-   melon: '103:0,16,1000,1,18'
-   cactus: '81:0,16,450,1,15'
-   sugarcane: '338:0,16,400,1,18'
-   wheat: '296:0,16,560,1,25'
- brewing:
-  name: '&cBrewing'
-  permission: 'customshop.main'
-  items:
-   fermentedspidereye: '376:0,4,1200,0,-'
-   ghasttear: '370:0,4,1200,0,-'
-   blazepower: '377:0,4,1200,0,-'
-   waterbottle: 373:0,16,800,0,-'
-   brewingstand: '379:0,1,250,0,-'
-   mushroom: '39:0,4,1200,0,-'
-   redmushroom: '40:0,4,1200,0,-'
-   magmacream: '378:0,4,1200,1,10'
-   goldencarrot: 396:0,4,1200,0,-'
-   glowstonedust: '348:0,4,1200,0,-'
-   pufferfish: '349:3,0,4,1200,0,-'
-   poisonouspotato: '394:0,4,1200,0,-'
-   lilypad: '111:0,4,1200,0,-'
-   sugar: '353:0,4,1200,0,-'
-   glisteringmelon: '382:0,4,1200,0,-'
- raiding:
-  name: '&cRaiding'
-  permission: 'customshop.main'
-  items:
-   tnt: '46:0,64,1500,0,-'
-   stickypiston: '29:0,1,150,0,-'
-   piston: '33:0,1,150,0,-'
-   lever: '69:0,1,25,0,-'
-   button: '77:0,1,25,0,-'
-   trappedchest: '146:0,32,500,0,-'
-   chest: '54:0,32,500,0,-'
-   redstonetorch: '76:0,8,200,0,-'
-   redstonerepeater: '356:0,1,100,0,-'
-   redstone: '331:0,8,100,0,-'
-   dispenser: '23:0,32,3200,0,-'
-   redstonecomparator: '404:0,1,100,0,-'
-   stonepressureplate: '70:0,1,50,0,-'
-   woodentrapdoor: '96:0,8,100,0,-'
-   cobweb: '30:0,16,800,0,-'
-   sponge: '19:0,64,1000,0,-'
-   glass: '20:0,32,100,0,-'
-   sand: '12:0,64,200,0,-'
-   redsand: '12:1,64,200,0,-'
-   ice: '79:0,16,400,0,-'
-   packedice: '174:0,64,600,0,-'
-   waterbucket: '326:0,1,100,0,-'
- blocks:
-  name: '&cBlocks'
-  permission: 'customshop.main'
-  items:
-   endstone: '121:0,64,250,1,1'
-   gravel: 13:0,64,175,1,1'
-   soulsand: 88:0,64,200,1,1'
-   netherbrick: '112:0,64,275,0,-'
-   mycelium: '110:0,32,250,0,-'
-   netherrack: '87:0,64,125,1,1'
-   glowstone: '89:0,64,350,1,1'
-   obsidian: '49:0,64,1000,0,-'
-   dirt: '3:0,64,150,1,1'
-   grass: '2:0,64,185,1,1'
-   cobblestone: '4:0,64,200,1,1'
-   stone: '1:0,64,275,1,1'
-   acacialog: '162:0,64,200,1,1'
-   oaklog: '17:0,64,200,1,1'
-   sprucelog: '17:1,64,200,1,1'
-   birchlog: '17:2,64,200,1,1'
-   junglelog: '17:3,64,200,1,1'
-   darkoakwood: '162:1,64,200,0,-'
-   sand: '12:0,64,200,0,-'
-   sandstone: '24:0,16,150,1,1'
-   chiseledsandstone: '24:1,16,175,0,-'
-   stonebricks: '98:0,64,350,0,-'
-   mossystonebricks: '98:1,64,350,0,-'
- blacksmith:
-  name: '&cBlacksmith'
-  permission: 'customshop.main'
-  items:
-   dpickaxe: '278:0,1,870,0,-'
-   dshovel: '277:0,1,290,0,-'
-   daxe: '279:0,1,870,0,-'
-   dsword: '276:0,1,500,0,-'
-   dboots: '313:0,1,1120,0,-'
-   dleggings: '312:0,1,1960,0,-'
-   dchestplate: '311:0,1,2250,0,-'
-   dhelmet: '310:0,1,1450,0,-'
-   chainboots: '305:0,1,560,0,-'
-   chainleggings: '304:0,1,580,0,-'
-   chainchestplate: '303:0,1,1125,0,-'
-   chathelmet: '302:0,1,725,0,-'
-   arrow: '262:0,64,200,1,1'
-   bow: '261:0,1,50,0,-'
-   fishingrod: '346:0,1,50,0,-'
- spawners:
-  name: '&cSpawners'
-  permission: 'customshop.spawners'
-  items:
-   test: '52:50,1,60000,0,-'
-   test1: '52:51,1,60000,0,-'
-   test2: '52:52,1,60000,0,-'
-   test3: '52:53,1,60000,0,-, buy:[/broadcast &e<player> just bought an Iron Golem Spawner!]'
